feat: remove scheduler in Reactor
When an action is sent to a reactor it is run on the current state and this process cannot be done from multiple threads

It is possible to make this process thread-safe by introducing scheduler, but this cause other problems

Those APIs tend to deliver their outputs on whatever thread is most convenient for them, and then it is your responsibility

ReactorKit makes you responsible for making sure to send actions on the main thread. If you may mutate states on a non-main thread, you must explicitly perform `.observe(on:)` in order to force it back on the main thread.