Expose mutation stream
This is different approach to the problem that Pulse is trying to solve.
I believe this is simpler and more intuitive logic for beginners. Since minimal Swift syntax is used, you do not need to know the Property Wrapper concept.

This revision emphasizes the foundational concept of event sourcing.
Mutation is the same as Event of EventSourcing. (Mutation will be renamed to Event in futher PR)
An event(mutation) represents a change in state.
When event are aggregated, they become a state.

It's important to recognize that not every event must directly influence the `state` model of Reactor.
For instance, events like 'alertMessageDidClose' and 'alertMessageDidChange' don't naturally fit within the state structure.
In the context of event sourcing—and by extension, ReactorKit—events may directly influence various "read" data models (the state of Reactor or UI components, such as UIKit or UIView, based on their intended purpose.